+++ runtime/devnet/src/pallets/utility.rs
@@ -1,0 +1,10 @@
+#![allow(unused_imports)]
+
+use crate::*;
+
+impl pallet_utility::Config for Runtime {
+	type RuntimeEvent = RuntimeEvent;
+	type RuntimeCall = RuntimeCall;
+	type PalletsOrigin = OriginCaller;
+	type WeightInfo = pallet_utility::weights::SubstrateWeight<Runtime>;
+}
